Make verification script buildserver friendly
Add better argument parsing for locked down files verification script
which accepts --whitelist <commit> and --import-gpg-keys arguments.
The default settings are supposed to work on the build server without
importing the gpg keys from the trusted_keys.pub file and running with a
hardcoded whitelist commit.

Make the CI workflow use these arguments as it is supposed to in
.github.