Deduplicate locked down files list
Remove the list of locked down files from locked_down_files.txt and
instead use the .github workflow as a single source of truth. This
requires some complicated parsing in the verification script as well as
a dependency from the verification script to the workflow YAML. These
are not ideal design choices however the alternative is to not have a
single source of truth for the locked down files as the github workflow
can not depend on an external file.